+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+# unreleased
+
+- (Label): uppercase text can be disabled setting `uppercase` prop to false.
+
 # v2.0.0-dev.5
 
 - (Select) added select and selectfield components that uses expanderBase and expanderSelect
 - (Tooltip) removed arrow
+
 
 # v2.0.0-dev.4